+++ src/main/battlecode/client/viewer/InfoPanel.java
@@ -78,12 +78,12 @@ public class InfoPanel extends JPanel {
   }
 
   public void updateDebugChanges(AbstractDrawObject<AbstractAnimation> robot,
-                                 int x, int y, int ore, int supply) {
+                                 int x, int y, double ore, double supply) {
     if (robot == null) {
       clear();
       robotID.setText("Ore: " + ore);
-      indicatorStrings[0].setText("Location: " + Integer.toString(x) + ", " + Integer.toString(y));
-      indicatorStrings[1].setText("Supply on ground: " + Integer.toString(supply));
+      indicatorStrings[0].setText("Location: " + Double.toString(x) + ", " + Integer.toString(y));
+      indicatorStrings[1].setText("Supply on ground: " + Double.toString(supply));
       //robotID.setText("No robot selected");
     }
     else {
